Various changes to prevent narrowing conversions: content/
Bug: 1216696
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2957544
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Nasko Oskov <nasko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#892335}

+++ content/browser/web_contents/web_contents_observer_browsertest.cc
@@ -234,7 +234,7 @@ class CookieTracker : public WebContentsObserver {
 
     ContextType context_type;
     GlobalRoutingID frame_id;
-    int navigation_id = -1;
+    int64_t navigation_id = -1;
 
     GURL url;
     GURL first_party_url;
